requirements1.txt
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
boto3==1.24.15
botocore==1.27.15
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
django-import-export==2.8.0
django-mathfilters==1.0.0
docopt==0.6.2
entrypoints==0.4
enum-compat==0.0.3
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
future==0.18.2
gunicorn==20.1.0
idna==3.3
ipykernel==6.10.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.8
jsonschema==4.6.0
jupyter-client==7.2.0
jupyter-core==4.10.0
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
nbclassic==0.4.3
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.1.6
notebook-shim==0.1.0
numpy==1.23.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
pipreqs==0.4.11
plotly==4.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.21.1
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
pydeck==0.3.0
PyGetWindow==0.0.9
Pygments==2.12.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.2.0
requests==2.27.1
retrying==1.3.3
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
tablib==3.2.1
terminado==0.13.3
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.2
traitlets==5.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
whitenoise==6.1.0
widgetsnbextension==3.6.1
wikipedia==1.4.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
